releaser: all packages are now checked after split of releasing branch instead of only project-base
