+++ utils/releaser/src/ReleaseWorker/ReleaseCandidate/CheckPackagesGithubActionsBuildsAfterSplitReleaseWorker.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\Releaser\ReleaseWorker\ReleaseCandidate;
+
+use PharIo\Version\Version;
+use Shopsys\Releaser\ReleaseWorker\AbstractCheckPackagesGithubActionsBuildsReleaseWorker;
+use Shopsys\Releaser\ReleaseWorker\AbstractShopsysReleaseWorker;
+use Shopsys\Releaser\Stage;
+
+final class CheckPackagesGithubActionsBuildsAfterSplitReleaseWorker extends AbstractCheckPackagesGithubActionsBuildsReleaseWorker
+{
+    /**
+     * @return string
+     */
+    public function getStage(): string
+    {
+        return Stage::RELEASE_CANDIDATE;
+    }
+
+    /**
+     * @return string
+     */
+    protected function getBranchName(): string
+    {
+        return $this->currentBranchName;
+    }
+
+    /**
+     * @param \PharIo\Version\Version $version
+     * @param string $initialBranchName
+     */
+    public function work(
+        Version $version,
+        string $initialBranchName = AbstractShopsysReleaseWorker::MAIN_BRANCH_NAME,
+    ): void {
+        $this->confirm(
+            sprintf(
+                'Confirm that current branch (%s) is split and all packages actions are finished.',
+                $this->getBranchName(),
+            ),
+        );
+
+        parent::work($version, $initialBranchName);
+    }
+}
