--- utils/releaser/src/ReleaseWorker/ReleaseCandidate/CheckProjectBaseBuild.php
@@ -1,48 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\Releaser\ReleaseWorker\ReleaseCandidate;
-
-use PharIo\Version\Version;
-use Shopsys\Releaser\ReleaseWorker\AbstractShopsysReleaseWorker;
-use Shopsys\Releaser\Stage;
-
-final class CheckProjectBaseBuild extends AbstractShopsysReleaseWorker
-{
-    /**
-     * @param \PharIo\Version\Version $version
-     * @param string $initialBranchName
-     * @return string
-     */
-    public function getDescription(
-        Version $version,
-        string $initialBranchName = AbstractShopsysReleaseWorker::MAIN_BRANCH_NAME,
-    ): string {
-        return '[Manually] Check project-base build on GitHub Actions';
-    }
-
-    /**
-     * @param \PharIo\Version\Version $version
-     * @param string $initialBranchName
-     */
-    public function work(
-        Version $version,
-        string $initialBranchName = AbstractShopsysReleaseWorker::MAIN_BRANCH_NAME,
-    ): void {
-        $this->symfonyStyle->note([
-            'check project-base build on GitHub Actions',
-            '  https://github.com/shopsys/project-base/actions/workflows/run-checks-tests.yaml',
-        ]);
-
-        $this->confirm('Confirm the build is ok');
-    }
-
-    /**
-     * @return string
-     */
-    public function getStage(): string
-    {
-        return Stage::RELEASE_CANDIDATE;
-    }
-}
